+++ src/components/frame/Extensions/Pages/MainExtensionPage/MainExtensionPage.tsx
@@ -1,8 +1,9 @@
 /* eslint-disable max-len */
-import React, { FC } from 'react';
+import React, { FC, useMemo } from 'react';
 import { useIntl } from 'react-intl';
 import ExtensionItem from '~common/Extensions/ExtensionItem';
 import PageTitle from '~common/Extensions/PageTitle';
+import TwoColumns from '~frame/Extensions/TwoColumns/TwoColumns';
 import { useExtensionsData, useMobile } from '~hooks';
 import { SpinnerLoader } from '~shared/Preloaders';
 
@@ -13,6 +14,11 @@ const MainExtensionPage: FC = () => {
   const { formatMessage } = useIntl();
   const isMobile = useMobile();
 
+  const allExtensions = useMemo(
+    () => [...availableExtensionsData, ...installedExtensionsData],
+    [availableExtensionsData, installedExtensionsData],
+  );
+
   if (loading) {
     return (
       <SpinnerLoader
@@ -21,8 +27,6 @@ const MainExtensionPage: FC = () => {
       />
     );
   }
-
-  const allExtensions = [...availableExtensionsData, ...installedExtensionsData];
 
   return (
     <div className="mb-6">
@@ -36,24 +40,27 @@ const MainExtensionPage: FC = () => {
       </div>
       <div className="flex lg:gap-[6.25rem] md:gap-12 mt-9">
         {/* @TODO: Add sidepanel */}
-        {!isMobile && <div>Sidepanel</div>}
-        <div>
-          <h4 className="text-xl font-semibold mb-6">{formatMessage({ id: 'extensionsPage.availableExtensions' })}</h4>
-          <h5 className="text-md font-semibold mb-4">{formatMessage({ id: 'extensionsPage.payments' })}</h5>
-          <ul className="pb-6 border-b border-gray-100">
-            {allExtensions.map((extension) => (
-              <li className="mb-6" key={extension.extensionId}>
-                <ExtensionItem
-                  title={extension.name}
-                  description={extension.descriptionShort}
-                  version={extension.availableVersion}
-                  icon={extension.icon}
-                  extensionId={extension.extensionId}
-                />
-              </li>
-            ))}
-          </ul>
-        </div>
+        <TwoColumns aside={<div>Sidepanel</div>}>
+          <div>
+            <h4 className="text-xl font-semibold mb-6">
+              {formatMessage({ id: 'extensionsPage.availableExtensions' })}
+            </h4>
+            <h5 className="text-md font-semibold mb-4">{formatMessage({ id: 'extensionsPage.payments' })}</h5>
+            <ul className="pb-6 border-b border-gray-100">
+              {allExtensions.map((extension) => (
+                <li className="mb-6" key={extension.extensionId}>
+                  <ExtensionItem
+                    title={extension.name}
+                    description={extension.descriptionShort}
+                    version={extension.availableVersion}
+                    icon={extension.icon}
+                    extensionId={extension.extensionId}
+                  />
+                </li>
+              ))}
+            </ul>
+          </div>
+        </TwoColumns>
       </div>
     </div>
   );
